requirements.txt
anyio==3.6.2
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.12.0
captcha==0.4
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
cryptography==39.0.2
cssselect2==0.7.0
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-filter==23.1
django-jazzmin==2.6.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-rest-framework==0.1.0
django-simple-captcha==0.5.17
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
executing==1.2.0
filelock==3.9.0
geojson==3.0.1
gmaps==0.9.0
google==3.0.0
googlemaps==4.10.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
ipykernel==6.21.1
ipython==8.10.0
ipywidgets==8.0.6
jedi==0.18.2
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyterlab-widgets==3.0.7
lxml==4.9.2
Markdown==3.4.3
matplotlib-inline==0.1.6
mysqlclient==2.1.1
nest-asyncio==1.5.6
oscrypto==1.3.0
outcome==1.2.0
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
pypdf==3.6.0
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-telegram-bot==20.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305
PyYAML==6.0
pyzmq==25.0.0
qrcode==7.4.2
reportlab==3.6.12
requests==2.28.2
rfc3986==1.5.0
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sqlparse==0.4.3
stack-data==0.6.2
svglib==1.5.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
uritools==4.0.1
urllib3==1.26.14
virtualenv==20.19.0
wcwidth==0.2.6
webencodings==0.5.1
widgetsnbextension==4.0.7
wsproto==1.2.0
xhtml2pdf==0.2.9
